requirements.txt
google-api-python-client==1.7.8
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
python-magic==0.4.15
hurry.filesize==0.9
pyrfc3339==1.1
